feat: reorganize client declaration
